backend/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
pymongo==4.15.4
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
